Move options to general type definitions folder
